docs: require docs on all public APIs
