+++ src/impl_.rs
@@ -1,3 +1,5 @@
+#![allow(missing_docs)]
+
 //! Internals of PyO3 which are accessed by code expanded from PyO3's procedural macros.
 //!
 //! Usage of any of these APIs in downstream code is implicitly acknowledging that these
