+++ src/inspect/mod.rs
@@ -1,4 +1,4 @@
-/// Runtime inspection of objects exposed to Python.
-///
-/// Tracking issue: <https://github.com/PyO3/pyo3/issues/2454>.
+//! Runtime inspection of objects exposed to Python.
+//!
+//! Tracking issue: <https://github.com/PyO3/pyo3/issues/2454>.
 pub mod types;
